Update files while rebaseing
