+++ Post/models.py
@@ -1,7 +1,6 @@
 from django.db import models
 from django.contrib.auth.models import User
 from django.utils import timezone
-from django.urls import reverse
 
 
 class Post(models.Model):
@@ -23,6 +22,3 @@ class Post(models.Model):
     @classmethod
     def all_posts(cls):
         return cls.objects.all()
-
-    def get_absolute_url(self):
-        return reverse('view posts')
